feat: add read grouping helper
