+++ custom_components/thessla_green_modbus/loader.py
@@ -15,6 +15,7 @@ from .registers import (
     get_all_registers,
     get_registers_by_function as _get_registers_by_function,
 )
+from .modbus_helpers import group_reads
 
 
 @lru_cache(maxsize=1)
@@ -34,26 +35,5 @@ def get_registers_by_function(function: str) -> Dict[str, int]:
     """Return mapping of register names to addresses for ``function``."""
 
     return {r.name: r.address for r in _get_registers_by_function(function)}
-
-
-def group_reads(addresses: Iterable[int], max_block_size: int = 64) -> List[Tuple[int, int]]:
-    """Group raw addresses into contiguous read blocks."""
-
-    sorted_addresses = sorted(set(addresses))
-    if not sorted_addresses:
-        return []
-
-    groups: List[Tuple[int, int]] = []
-    start = prev = sorted_addresses[0]
-    for addr in sorted_addresses[1:]:
-        if addr == prev + 1 and (addr - start + 1) <= max_block_size:
-            prev = addr
-            continue
-        groups.append((start, prev - start + 1))
-        start = prev = addr
-    groups.append((start, prev - start + 1))
-    return groups
-
-
 __all__ = ["Register", "get_register_definition", "get_registers_by_function", "group_reads"]
